wikipedia_parser.articles: init
Move articles json processing to wikipedia_parser.articles
